Fix af24192b0f (support for struct_time in utils.str.timestamp).

+++ src/conf.py
@@ -374,7 +374,7 @@ registerChannelValue(supybot.reply.format, 'time',
 def timestamp(t):
     if t is None:
         t = time.time()
-    elif isinstance(t, int):
+    elif isinstance(t, float):
         t = time.localtime(t)
     format = get(supybot.reply.format.time, dynamic.channel)
     return time.strftime(format, t)
